Fix more AppVeyor warnings

+++ src/video/gl/gl_lightmap.hpp
@@ -51,7 +51,7 @@ private:
   int m_lightmap_height;
   float m_lightmap_uv_right;
   float m_lightmap_uv_bottom;
-  GLfloat m_old_viewport[4]; //holds vieport before redefining in start_draw - returned from glGet
+  GLint m_old_viewport[4]; //holds vieport before redefining in start_draw - returned from glGet
 
 private:
   GLLightmap(const GLLightmap&);
